Move MQTT platforms under the component
