+++ homeassistant/components/mqtt/binary_sensor.py
