+++ homeassistant/components/mqtt/fan.py
