+++ homeassistant/components/mqtt/sensor.py
