+++ homeassistant/components/mqtt/switch.py
