+++ tests/components/mqtt/test_alarm_control_panel.py
