+++ tests/components/mqtt/test_binary_sensor.py
