+++ tests/components/mqtt/test_climate.py
