+++ tests/components/mqtt/test_fan.py
