+++ tests/components/mqtt/test_lock.py
