+++ tests/components/mqtt/test_switch.py
